Safari-611.1.10/LayoutTests/fast/css/getComputedStyle/empty-update-without-renderer-expected.txt

When updating the tree, the style needs to be invalidated when the :empty state, even if there is no renderer.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyle(document.getElementById("target")).color is "rgb(0, 0, 0)"
PASS getComputedStyle(document.getElementById("target")).color is "rgb(1, 2, 3)"
PASS successfullyParsed is true

TEST COMPLETE
